add transaction process bar to all apps
